refactor(utils.api-harmonization): pagination resolver out of the models
`framework/src/utils/models` is for models, and the resolver that turns a query's
`limit`, `offset` and `page` into the window to fetch had landed there. It moves to
`Utils.Pagination.resolvePagination`, next to the other helpers the API side
already reaches for as `Utils.Date`, `Utils.Price` and `Utils.Auth`.

Not to `@o2s/utils.frontend`, which the review suggested: only the blocks' NestJS
services call this, and that package depends on `@o2s/ui` and React, so a server
importing it would drag the component library into the API. The query shape stays
a model in `@o2s/framework` — `Models.Pagination.PaginatedQuery` — while the
resolving lives with the helpers.

The fallback page size moves with it and stops being a single row. Four of the five
blocks fell back to `limit: 1` when neither the query nor the CMS config named one.
That was inherited, and reachable: `pagination` is optional in the CMS block
models, so an entry without it rendered a one-row list. `DEFAULT_LIMIT` (10) is the
shared fallback now, and the product list keeps its own 12 for its three-column
grid.

The spec moves along with the code, and the package gets the test project it never
had. The blocks that now import it say so in their dependencies.
