+++ packages/blocks/billing/invoice-list/package.json
@@ -38,6 +38,7 @@
     "dependencies": {
         "@o2s/configs.integrations": "*",
         "@o2s/ui": "*",
+        "@o2s/utils.api-harmonization": "*",
         "@o2s/utils.logger": "*",
         "express": "^5.2.1",
         "intl-messageformat": "^11.1.0",
