+++ packages/blocks/notifications/notification-list/src/api-harmonization/notification-list.service.ts
@@ -2,16 +2,16 @@ import { Injectable } from '@nestjs/common';
 import { CMS, Notifications } from '@o2s/configs.integrations';
 import { Observable, concatMap, forkJoin, map } from 'rxjs';
 
+import { Utils } from '@o2s/utils.api-harmonization';
+
 import { AppHeaders, HeaderName } from '@o2s/framework/headers';
-import { Auth, Models } from '@o2s/framework/modules';
+import { Auth } from '@o2s/framework/modules';
 
 import { mapNotificationList } from './notification-list.mapper';
 import { NotificationListBlock } from './notification-list.model';
 import { GetNotificationListBlockQuery } from './notification-list.request';
 
 const H = HeaderName;
-
-const DEFAULT_LIMIT = 1;
 
 @Injectable()
 export class NotificationListService {
@@ -36,9 +36,8 @@ export class NotificationListService {
             concatMap(([cms]) => {
                 // `page` is a URL concern and is consumed here, so it never reaches the domain module.
                 const { page: _page, ...notificationQuery } = query;
-                const { limit, offset } = Models.Pagination.resolvePagination(query, {
+                const { limit, offset } = Utils.Pagination.resolvePagination(query, {
                     cmsLimit: cms.pagination?.limit,
-                    defaultLimit: DEFAULT_LIMIT,
                 });
 
                 return this.notificationService
