+++ packages/blocks/orders/order-list/src/api-harmonization/order-list.service.ts
@@ -2,16 +2,16 @@ import { Injectable } from '@nestjs/common';
 import { CMS, Orders } from '@o2s/configs.integrations';
 import { Observable, concatMap, forkJoin, map } from 'rxjs';
 
+import { Utils } from '@o2s/utils.api-harmonization';
+
 import { AppHeaders, HeaderName } from '@o2s/framework/headers';
-import { Auth, Models } from '@o2s/framework/modules';
+import { Auth } from '@o2s/framework/modules';
 
 import { mapOrderList } from './order-list.mapper';
 import { OrderListBlock } from './order-list.model';
 import { GetOrderListBlockQuery } from './order-list.request';
 
 const H = HeaderName;
-
-const DEFAULT_LIMIT = 1;
 
 @Injectable()
 export class OrderListService {
@@ -33,9 +33,8 @@ export class OrderListService {
             concatMap(([cms]) => {
                 // `page` is a URL concern and is consumed here, so it never reaches the domain module.
                 const { page: _page, ...orderQuery } = query;
-                const { limit, offset } = Models.Pagination.resolvePagination(query, {
+                const { limit, offset } = Utils.Pagination.resolvePagination(query, {
                     cmsLimit: cms.pagination?.limit,
-                    defaultLimit: DEFAULT_LIMIT,
                 });
 
                 return this.orderService
