+++ packages/blocks/products/product-list/src/api-harmonization/product-list.service.ts
@@ -2,8 +2,10 @@ import { Injectable } from '@nestjs/common';
 import { CMS, Products } from '@o2s/configs.integrations';
 import { Observable, concatMap, forkJoin, map } from 'rxjs';
 
+import { Utils } from '@o2s/utils.api-harmonization';
+
 import { AppHeaders, HeaderName } from '@o2s/framework/headers';
-import { Auth, Models } from '@o2s/framework/modules';
+import { Auth } from '@o2s/framework/modules';
 
 import { mapProductList } from './product-list.mapper';
 import { ProductListBlock } from './product-list.model';
@@ -11,6 +13,7 @@ import { GetProductListBlockQuery } from './product-list.request';
 
 const H = HeaderName;
 
+/** A grid of three columns, so its own fallback rather than the shared one. */
 const DEFAULT_LIMIT = 12;
 
 @Injectable()
@@ -32,7 +35,7 @@ export class ProductListService {
             concatMap(([cms]) => {
                 // `page` is a URL concern and is consumed here, so it never reaches the products module.
                 const { page: _page, ...productQuery } = query;
-                const { limit, offset } = Models.Pagination.resolvePagination(query, {
+                const { limit, offset } = Utils.Pagination.resolvePagination(query, {
                     cmsLimit: cms.pagination?.limit,
                     defaultLimit: DEFAULT_LIMIT,
                 });
