+++ packages/blocks/support/ticket-list/src/api-harmonization/ticket-list.service.ts
@@ -2,16 +2,16 @@ import { Injectable } from '@nestjs/common';
 import { CMS, Tickets } from '@o2s/configs.integrations';
 import { Observable, concatMap, forkJoin, map } from 'rxjs';
 
+import { Utils } from '@o2s/utils.api-harmonization';
+
 import { AppHeaders, HeaderName } from '@o2s/framework/headers';
-import { Auth, Models } from '@o2s/framework/modules';
+import { Auth } from '@o2s/framework/modules';
 
 import { mapTicketList } from './ticket-list.mapper';
 import { TicketListBlock } from './ticket-list.model';
 import { GetTicketListBlockQuery } from './ticket-list.request';
 
 const H = HeaderName;
-
-const DEFAULT_LIMIT = 1;
 
 @Injectable()
 export class TicketListService {
@@ -33,9 +33,8 @@ export class TicketListService {
             concatMap(([cms]) => {
                 // `page` is a URL concern and is consumed here, so it never reaches the tickets module.
                 const { page: _page, ...ticketQuery } = query;
-                const { limit, offset } = Models.Pagination.resolvePagination(query, {
+                const { limit, offset } = Utils.Pagination.resolvePagination(query, {
                     cmsLimit: cms.pagination?.limit,
-                    defaultLimit: DEFAULT_LIMIT,
                 });
 
                 return this.ticketService
