+++ packages/utils/api-harmonization/package.json
@@ -28,6 +28,7 @@
     "scripts": {
         "build": "tsc --preserveWatchOutput && tsc-alias",
         "lint": "tsc --noEmit && eslint \"src/**/*.{ts,tsx}\" --fix --max-warnings=0",
+        "test": "vitest run",
         "format": "prettier --write \"src/**/*.{ts,tsx}\""
     },
     "devDependencies": {
@@ -35,6 +36,7 @@
         "@o2s/lint-staged-config": "*",
         "@o2s/prettier-config": "*",
         "@o2s/typescript-config": "*",
+        "@o2s/vitest-config": "*",
         "concurrently": "^9.2.1",
         "eslint": "^9.39.4",
         "prettier": "^3.8.1",
