+++ packages/utils/api-harmonization/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * as Date from './date';
 export * as Price from './price';
 export * as Auth from './auth';
+export * as Pagination from './pagination';
