+++ packages/utils/api-harmonization/vitest.config.mjs
@@ -1,0 +1,3 @@
+import { config } from '@o2s/vitest-config/api';
+
+export default config;
